+++ changelog.md
@@ -4,6 +4,8 @@
 
 ### TODO
 
+- make models on angular side for message, room and user
+- angular functionality for deleting single message
 - add push notifcation
 - show recent chats
 - send message after request accept
@@ -13,6 +15,8 @@
 ### Added
 
 - Added prettier eslint
+- add nested schema for message
+- add api for deleting single message
 - fetch last 100 messages in a chat
 
 ### Fixed